retrosheet_fetcher/requirements.txt
requests
functions-framework
google-cloud-storage